requirements.txt
DJANGO==3.2.7
pillow
djangorestframework==3.12.4